Fix warning

+++ src/Analyses/ATLAS_2011_I919017.cc
@@ -29,7 +29,7 @@ namespace Rivet {
 
     // For annulus area kludge
     inline double calcrweight(const Jet&j, const Particle& p) {
-      int nBins_r = 26;
+      size_t nBins_r = 26;
       double bins_r[] = { 0.00, 0.01, 0.02, 0.03, 0.04, 0.05, 0.06, 0.07, 0.08, 0.09, 0.10,
                           0.12, 0.14, 0.16, 0.18, 0.20, 0.22, 0.24, 0.26, 0.28, 0.30,
                           0.35, 0.40, 0.45, 0.50, 0.55, 0.60 };
